+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 
 * `#3 <https://github.com/xitrum-framework/scala-xgettext/issues/3>`_
   Make i18n method names configurable
+* `#4 <https://github.com/xitrum-framework/scala-xgettext/issues/4>`_
+  Poedit reports "invalid control sequence" for key "don\'t have to"
 
 1.1: